Chore(.security): move to annotation package
