+++ src/main/java/com/newbarams/ajaja/module/plan/presentation/PlanController.java
@@ -21,8 +21,8 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.newbarams.ajaja.global.common.AjajaResponse;
-import com.newbarams.ajaja.global.security.common.UserId;
-import com.newbarams.ajaja.global.security.jwt.util.JwtParser;
+import com.newbarams.ajaja.global.security.annotation.UserId;
+import com.newbarams.ajaja.global.security.jwt.JwtParser;
 import com.newbarams.ajaja.global.util.BearerUtil;
 import com.newbarams.ajaja.module.ajaja.application.SwitchAjajaService;
 import com.newbarams.ajaja.module.plan.application.CreatePlanService;
