+++ src/main/java/com/newbarams/ajaja/module/remind/adapter/in/web/SendTestRemindController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.newbarams.ajaja.global.common.AjajaResponse;
-import com.newbarams.ajaja.global.security.common.UserId;
+import com.newbarams.ajaja.global.security.annotation.UserId;
 import com.newbarams.ajaja.module.remind.application.port.in.SendTestRemindUseCase;
 
 import lombok.RequiredArgsConstructor;
