+++ src/main/java/com/newbarams/ajaja/module/user/adapter/in/web/ChangeRemindTypeController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.newbarams.ajaja.global.common.AjajaResponse;
-import com.newbarams.ajaja.global.security.common.UserId;
+import com.newbarams.ajaja.global.security.annotation.UserId;
 import com.newbarams.ajaja.module.user.application.port.in.ChangeRemindTypeUseCase;
 import com.newbarams.ajaja.module.user.dto.UserRequest;
