+++ src/main/java/com/newbarams/ajaja/module/user/adapter/in/web/LogoutController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.newbarams.ajaja.global.security.common.UserId;
+import com.newbarams.ajaja.global.security.annotation.UserId;
 import com.newbarams.ajaja.module.user.application.port.in.LogoutUseCase;
 
 import lombok.RequiredArgsConstructor;
