+++ src/main/java/com/newbarams/ajaja/module/user/adapter/in/web/SendVerificationController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.newbarams.ajaja.global.security.common.UserId;
+import com.newbarams.ajaja.global.security.annotation.UserId;
 import com.newbarams.ajaja.module.user.application.port.in.SendVerificationEmailUseCase;
 import com.newbarams.ajaja.module.user.dto.UserRequest;
